+++ packages/sdk-react/src/components/composite/index.ts
@@ -43,5 +43,4 @@ export * from './MobileControlSheet';
 export * from './MobilePanel';
 export * from './ReactionPicker';
 export * from './RecordingControls';
-export * from './PictureInPicture';
 export * from './WhatsNewDialog';